requirements.txt
streamlit>=1.37.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
numpy>=1.26.0
scikit-learn>=1.4.0
rank-bm25>=0.2.2
langchain-community>=0.2.7
pypdf>=4.2.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
openai>=1.40.0
langdetect>=1.0.9
transformers>=4.30.0
torch>=2.0.0
deep-translator>=1.11.0
